+++ src/snarky/web/getEfficientNumWorkers.js
