+++ src/snarky/web/worker_run.js
