+++ src/snarky/wrapper.web.js
@@ -1,4 +1,4 @@
-import { initSnarkyJS } from '../chrome_bindings/plonk_init.js';
+import { initSnarkyJS } from './web/plonk_init.js';
 
 export { getSnarky, getWasm, snarkyReady, shutdown };
